Add empty pages

+++ 2/pages/convert.sh
@@ -10,18 +10,38 @@ if ! exiftool -ver &>/dev/null; then
   exit 1
 fi
 
+if ! [ -e 776x1053/ ]; then
+  >&2 echo 'directory 776x1053/ not found; please run `bash getpages.sh` first'
+  exit 1
+fi
+
 md() { mkdir -p "$1" && cd "$1"; }
 
 ___jpeg() {  # remove metadata and compress slightly (images become about 90% in size)
   (md "$3"
   [ -s "$1.jpeg" ] || rm -f "$1.jpeg"  # if empty
   [ -e "$1.jpeg" ] || {
+    if   [ "$1" = "605" ]; then ___jpeg "empty-odd"  "$2" "."; return
+    elif [ "$1" = "606" ]; then ___jpeg "empty-even" "$2" "."; return
+    fi
     # remove metadata, except the color profile (which is only 4KB, but might cause color issues?)
-    # then compress to jpeg twice (only twice to not alter the color quality noticeably)
-    exiftool -q -all= --ICC_Profile:all "$2$1.jpg" -o "$1...jpeg" &&
-    convert "$1...jpeg" "$1..jpeg" &&
-    convert "$1..jpeg" "$1.jpeg" &&
-    true; } &&
+    # then compress the first two page twice (only twice to not alter the color quality noticeably)
+    # but the rest don't compress better, thus kept. then provide a numeric name for the empty pages.
+    exiftool -q -all= --ICC_Profile:all "$2$1.jpg" -o "$1...jpeg" && {
+      case "$1" in
+      1|2)
+        convert "$1...jpeg" "$1..jpeg" &&
+        convert "$1..jpeg" "$1.jpeg" ;;
+      empty-odd)
+        mv "$1...jpeg" "$1.jpeg" &&
+        { ln "$1.jpeg" "605.jpeg" || cp -f "$1.jpeg" "605.jpeg"; } ;;
+      empty-even)
+        mv "$1...jpeg" "$1.jpeg" &&
+        { ln "$1.jpeg" "606.jpeg" || cp -f "$1.jpeg" "606.jpeg"; } ;;
+      *)
+        mv "$1...jpeg" "$1.jpeg" ;;
+      esac
+    }; } &&
     rm -f "$1...jpeg" "$1..jpeg" ||
       goterror "$1.jpeg"
   echo -n "$1"/;)  # auto `cd -` because of the subshell
@@ -31,27 +51,46 @@ ___webp() {  # remove metadata and compress to webp (images become about 50% in 
   (md "$3"
   [ -s "$1.webp" ] || rm -f "$1.webp"  # if empty
   [ -e "$1.webp" ] || {
+    if   [ "$1" = "605" ]; then ___webp "empty-odd"  "$2" "."; return
+    elif [ "$1" = "606" ]; then ___webp "empty-even" "$2" "."; return
+    fi
     # remove metadata, except the color profile (which is only 4KB, but might cause color issues?)
     # then compress to webp twice (once is not enough to reduce its size)
     exiftool -q -all= --ICC_Profile:all "$2$1.jpg" -o "$1.jpg" &&
     convert "$1.jpg" "$1.WEBP" &&
-    convert "$1.WEBP" "$1.webp"; } &&
+    convert "$1.WEBP" "$1.webp" && {
+      case "$1" in
+      empty-odd)
+        ln "$1.webp" "605.webp" || cp -f "$1.webp" "605.webp" ;;
+      empty-even)
+        ln "$1.webp" "606.webp" || cp -f "$1.webp" "606.webp" ;;
+      esac
+    }; } &&
     rm -f "$1.jpg" "$1.WEBP" ||
       goterror "$1.webp"
   echo -n "$1"/;)  # auto `cd -` because of the subshell
 }
  
-___avif() {  # remove metadata and compress to avif (images become about 50% in size)
+___avif() {  # remove metadata and compress to avif (images become about 37% in size)
   (md "$3"
   [ -s "$1.avif" ] || rm -f "$1.avif"  # if empty
   [ -e "$1.avif" ] || {
+    if   [ "$1" = "605" ]; then ___avif "empty-odd"  "$2" "."; return
+    elif [ "$1" = "606" ]; then ___avif "empty-even" "$2" "."; return
+    fi
     # remove metadata, except the color profile (which is only 4KB, but might cause color issues?)
     # then compress to avif thrice
     exiftool -q -all= --ICC_Profile:all "$2$1.jpg" -o "$1.jpg" &&
     convert "$1.jpg" "$1...avif" &&
     convert "$1...avif" "$1..avif" &&
-    convert "$1..avif" "$1.avif" &&
-    true; } &&
+    convert "$1..avif" "$1.avif" && {
+      case "$1" in
+      empty-odd)
+        ln "$1.avif" "605.avif" || cp -f "$1.avif" "605.avif" ;;
+      empty-even)
+        ln "$1.avif" "606.avif" || cp -f "$1.avif" "606.avif" ;;
+      esac
+    }; } &&
     rm -f "$1.jpg" "$1...avif" "$1..avif" ||
       goterror "$1.avif"
   echo -n "$1"/;)  # auto `cd -` because of the subshell
@@ -69,7 +108,7 @@ loop() {
   # $2 is the number of concurrent processes
   # the rest are passed to function after the page number
   printf "\n%s\n" $1
-  for i in {1..604}; do
+  for i in {1..606}; do  # pages 605 and 606 are empty-odd & empty-even, repsectively
     $1 $i "${@:3}" &
     [ $((i % $2)) -eq 0 ] && wait &>/dev/null
     haserror && break
@@ -93,7 +132,7 @@ goterror() { echo -n E >> "$t"; [ -n "$1" ] && rm -rf "$1"; }
 
 # loop _light_jpeg $N
 
-# loop _light_webp $N
+loop _light_webp $N
 
 # loop _light_avif $N
